Fix edge flip
This changes the constraints in the edge flip routine to preserve the
adjacency of the computed quadrilateral to its exterior, which might not
all be -1 as was the case before. Also makes triangulating skeletons
return the transformation, which can then be modified some more, rather
than computing the result outright.