Add test case that makes convex hull fail
One of the input points is coplanar with the starting triangle. When the
initial visibility graph gets computed, it's record as 0.0 and so gets
skipped on every subsequent step, even though it becomes visible to
faces that are added later. We need to change how visibility is computed
to fix this.