Change how polytopal.from_simplicial works
Before, we would use the max element of the simplexes to decide how many
vertices there were, so if we got a simplicial complex that looked like
`[[14, 15, 16], [17, 16, 15]]` we would assume there were 18 vertices.
Now each vertex is given its own ID which cuts down the return size if
we want to extract a polytopal complex that happens to start at a
different index.